[#189] refactor: 스크롤방지 코드위치 변경
